fix penalty

+++ OpenNos.DAL.Interface/IPenaltyLogDAO.cs
@@ -26,7 +26,7 @@ namespace OpenNos.DAL.Interface
         IEnumerable<PenaltyLogDTO> LoadByAccount(long accountId);              
         SaveResult InsertOrUpdate(ref PenaltyLogDTO log);
         IEnumerable<PenaltyLogDTO> LoadAll();
-        PenaltyLogDTO LoadById(long relId);
+        PenaltyLogDTO LoadById(int relId);
 
         #endregion
     }
